+++ src/Json.Schema.ToDotNet.UnitTests/DataModelGeneratorTests.cs
@@ -229,245 +229,11 @@ namespace Microsoft.Json.Schema.ToDotNet.UnitTests
         [Fact(DisplayName = "DataModelGenerator generates XML comments for properties whose property type is ref")]
         public void GeneratesXmlCommentsForPropertiesWhosePropertyTypeIsRef()
         {
-            const string Schema =
-@"{
-  ""type"": ""object"",
-  ""description"": ""Describes a console window."",
-  ""properties"": {
-    ""foregroundColor"": {
-      ""$ref"": ""#/definitions/color"",
-      ""description"": ""The color of the text on the screen.""
-    },
-    ""backgroundColor"": {
-      ""$ref"": ""#/definitions/color"",
-      ""description"": ""The color of the screen background.""
-    },
-  },
-  ""definitions"": {
-    ""color"": {
-      ""type"": ""object"",
-      ""description"": ""Describes a color with R, G, and B components."",
-      ""properties"": {
-        ""red"": {
-          ""type"": ""integer"",
-          ""description"": ""The value of the R component.""
-        },
-        ""green"": {
-          ""type"": ""integer"",
-          ""description"": ""The value of the G component.""
-        },
-        ""blue"": {
-          ""type"": ""integer"",
-          ""description"": ""The value of the B component.""
-        }
-      }
-    }
-  }
-}";
-
-            const string ExpectedRootClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
-
-namespace N
-{
-    /// <summary>
-    /// Describes a console window.
-    /// </summary>
-    [DataContract]
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public partial class ConsoleWindow
-    {
-        public static IEqualityComparer<ConsoleWindow> ValueComparer => ConsoleWindowEqualityComparer.Instance;
-
-        public bool ValueEquals(ConsoleWindow other) => ValueComparer.Equals(this, other);
-        public int ValueGetHashCode() => ValueComparer.GetHashCode(this);
-
-        /// <summary>
-        /// The color of the text on the screen.
-        /// </summary>
-        [DataMember(Name = ""foregroundColor"", IsRequired = false, EmitDefaultValue = false)]
-        public Color ForegroundColor { get; set; }
-
-        /// <summary>
-        /// The color of the screen background.
-        /// </summary>
-        [DataMember(Name = ""backgroundColor"", IsRequired = false, EmitDefaultValue = false)]
-        public Color BackgroundColor { get; set; }
-    }
-}";
-            const string ExpectedRootComparerClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-
-namespace N
-{
-    /// <summary>
-    /// Defines methods to support the comparison of objects of type ConsoleWindow for equality.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    internal sealed class ConsoleWindowEqualityComparer : IEqualityComparer<ConsoleWindow>
-    {
-        internal static readonly ConsoleWindowEqualityComparer Instance = new ConsoleWindowEqualityComparer();
-
-        public bool Equals(ConsoleWindow left, ConsoleWindow right)
-        {
-            if (ReferenceEquals(left, right))
-            {
-                return true;
-            }
-
-            if (ReferenceEquals(left, null) || ReferenceEquals(right, null))
-            {
-                return false;
-            }
-
-            if (!Color.ValueComparer.Equals(left.ForegroundColor, right.ForegroundColor))
-            {
-                return false;
-            }
-
-            if (!Color.ValueComparer.Equals(left.BackgroundColor, right.BackgroundColor))
-            {
-                return false;
-            }
-
-            return true;
-        }
-
-        public int GetHashCode(ConsoleWindow obj)
-        {
-            if (ReferenceEquals(obj, null))
-            {
-                return 0;
-            }
-
-            int result = 17;
-            unchecked
-            {
-                if (obj.ForegroundColor != null)
-                {
-                    result = (result * 31) + obj.ForegroundColor.ValueGetHashCode();
-                }
-
-                if (obj.BackgroundColor != null)
-                {
-                    result = (result * 31) + obj.BackgroundColor.ValueGetHashCode();
-                }
-            }
-
-            return result;
-        }
-    }
-}";
-
-            const string ExpectedColorClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
-
-namespace N
-{
-    /// <summary>
-    /// Describes a color with R, G, and B components.
-    /// </summary>
-    [DataContract]
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public partial class Color
-    {
-        public static IEqualityComparer<Color> ValueComparer => ColorEqualityComparer.Instance;
-
-        public bool ValueEquals(Color other) => ValueComparer.Equals(this, other);
-        public int ValueGetHashCode() => ValueComparer.GetHashCode(this);
-
-        /// <summary>
-        /// The value of the R component.
-        /// </summary>
-        [DataMember(Name = ""red"", IsRequired = false, EmitDefaultValue = false)]
-        public int Red { get; set; }
-
-        /// <summary>
-        /// The value of the G component.
-        /// </summary>
-        [DataMember(Name = ""green"", IsRequired = false, EmitDefaultValue = false)]
-        public int Green { get; set; }
-
-        /// <summary>
-        /// The value of the B component.
-        /// </summary>
-        [DataMember(Name = ""blue"", IsRequired = false, EmitDefaultValue = false)]
-        public int Blue { get; set; }
-    }
-}";
-
-            const string ExpectedColorComparerClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-
-namespace N
-{
-    /// <summary>
-    /// Defines methods to support the comparison of objects of type Color for equality.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    internal sealed class ColorEqualityComparer : IEqualityComparer<Color>
-    {
-        internal static readonly ColorEqualityComparer Instance = new ColorEqualityComparer();
-
-        public bool Equals(Color left, Color right)
-        {
-            if (ReferenceEquals(left, right))
-            {
-                return true;
-            }
-
-            if (ReferenceEquals(left, null) || ReferenceEquals(right, null))
-            {
-                return false;
-            }
-
-            if (left.Red != right.Red)
-            {
-                return false;
-            }
-
-            if (left.Green != right.Green)
-            {
-                return false;
-            }
-
-            if (left.Blue != right.Blue)
-            {
-                return false;
-            }
-
-            return true;
-        }
-
-        public int GetHashCode(Color obj)
-        {
-            if (ReferenceEquals(obj, null))
-            {
-                return 0;
-            }
-
-            int result = 17;
-            unchecked
-            {
-                result = (result * 31) + obj.Red.GetHashCode();
-                result = (result * 31) + obj.Green.GetHashCode();
-                result = (result * 31) + obj.Blue.GetHashCode();
-            }
-
-            return result;
-        }
-    }
-}";
+            string Schema = TestUtil.ReadTestInputFile(this.GetType().Name, "Schema.json");
+            string ExpectedRootClass = TestUtil.ReadTestInputFile(this.GetType().Name, "ExpectedRootClass.cs");
+            string ExpectedRootComparerClass = TestUtil.ReadTestInputFile(this.GetType().Name, "ExpectedRootComparerClass.cs");
+            string ExpectedColorClass = TestUtil.ReadTestInputFile(this.GetType().Name, "ExpectedColorClass.cs");
+            string ExpectedColorComparerClass = TestUtil.ReadTestInputFile(this.GetType().Name, "ExpectedColorComparerClass.cs");
 
             _settings.GenerateEqualityComparers = true;
             _settings.RootClassName = "ConsoleWindow";
